requirements.txt
absl-py==2.0.0
aiohttp
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
async-timeout==4.0.3
attrs==23.1.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
colorama==0.4.6
contourpy
cycler==0.12.1
filelock==3.12.4
fonttools==4.43.1
frozenlist
fsspec==2023.9.2
google-auth==2.23.3
google-auth-oauthlib==1.0.0
grpcio
hydra-core==1.3.2
idna==3.4
Jinja2
joblib==1.3.2
kiwisolver
lightning-utilities
Markdown==3.5
MarkupSafe==2.1.3
matplotlib
mpmath
multidict
networkx
numpy
oauthlib==3.2.2
omegaconf==2.3.0
packaging==23.2
Pillow
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytorch-lightning
PyYAML
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn
scipy
six==1.16.0
sympy
tensorboard
tensorboard-data-server
threadpoolctl==3.2.0
torch
torchmetrics
torchvision
tqdm==4.66.1
typing_extensions
urllib3==2.0.6
Werkzeug==3.0.0
yarl
